+++ src/Cmd/Install.hs
@@ -31,9 +31,9 @@ import RpmBuild
 -- FIXME allow building an srpm
 -- FIXME ExistingStrategy option
 installCmd :: Bool -> Bool -> Maybe Branch -> Maybe ForceShort -> [BCond]
-           -> Bool -> Bool -> Bool -> Select -> (Maybe Branch,[String])
-           -> IO ()
-installCmd quiet recurse mfrom mforceshort bconds reinstall nobuild nobuilddeps select (mbr, pkgs) = do
+           -> Bool -> Bool -> Bool -> Select -> Maybe ExistingStrategy
+           -> (Maybe Branch,[String]) -> IO ()
+installCmd quiet recurse mfrom mforceshort bconds reinstall nobuild nobuilddeps select mexisting (mbr, pkgs) = do
   when (recurse && isShortCircuit mforceshort) $
     error' "cannot use --recurse and --shortcircuit"
   withPackagesMaybeBranch (boolHeader (recurse || length pkgs > 1)) True Nothing installPkg (mbr, pkgs)
@@ -76,7 +76,7 @@ installCmd quiet recurse mfrom mforceshort bconds reinstall nobuild nobuilddeps 
                   mpkgdir <- lookForPkgDir rbr ".." dep
                   case mpkgdir of
                     Nothing -> putStrLn $ dep +-+ "not known"
-                    Just pkgdir -> installCmd quiet recurse mfrom mforceshort bconds reinstall nobuild nobuilddeps select (mbr, [pkgdir]) >> putNewLn
+                    Just pkgdir -> installCmd quiet recurse mfrom mforceshort bconds reinstall nobuild nobuilddeps select mexisting (mbr, [pkgdir]) >> putNewLn
                 -- FIXME option to enable/disable installing missing deps
                 -- FIXME --skip-missing-deps or prompt
               else installDeps True spec
@@ -88,7 +88,7 @@ installCmd quiet recurse mfrom mforceshort bconds reinstall nobuild nobuilddeps 
           unless (isShortCircuit mforceshort') $ do
             let nvras = rpmsToNVRAs rpms
                 -- FIXME: prefix = fromMaybe (nvrName nvr) mprefix
-            decided <- decideRPMs No False Nothing select (unPackage pkg) nvras
+            decided <- decideRPMs No False mexisting select (unPackage pkg) nvras
             -- FIXME dryrun and debug
             -- FIXME return Bool?
             installRPMs False False Nothing No $ groupOnArch "RPMS" decided