+++ src/Cmd/ReviewPackage.hs
@@ -119,7 +119,7 @@ doInteractiveReview importsrpm mspec srpm = do
   void $ cmdBool "rpmlint" ["."] -- FIXME $ spec:srpm:rpms
   spec <- maybe findSpecfile return mspec
   whenM (yesNoDefault importsrpm "Install packages locally") $ do
-    installCmd False False Nothing Nothing [] False True True selectDefault (Nothing,[])
+    installCmd False False Nothing Nothing [] False True True selectDefault Nothing (Nothing,[])
     rpms <- cmdLines "rpmspec" ["-q", "--rpms", "--qf", "%{name}\n", spec]
     whenM (yesNoDefault importsrpm "Rpmlint installed packages") $ do
        (_ok, out, err) <- cmdFull "rpmlint" ("-i" : rpms) ""