+++ src/Main.hs
@@ -8,7 +8,7 @@ import Distribution.Fedora.Branch (partitionBranches, readBranch)
 #if !MIN_VERSION_simple_cmd_args(0,1,7)
 import Options.Applicative (maybeReader, ReadM)
 #endif
-import SelectRPMs (selectRpmsOption)
+import SelectRPMs (existingStrategyOption, selectRpmsOptions)
 import SimpleCmdArgs
 
 -- commands
@@ -266,7 +266,8 @@ main = do
       <*> switchWith 'r' "reinstall" "reinstall rpms"
       <*> switchLongWith "no-build" "do not (re-)build (install built rpms)"
       <*> switchLongWith "ignore-builddeps" "do not install builddeps"
-      <*> selectRpmsOption
+      <*> selectRpmsOptions
+      <*> optional existingStrategyOption
       <*> maybeBranchPackages False
     , Subcommand "not-installed" "Packages not installed locally" $
       notInstalledCmd