+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - 'create-review': --force to create a new "duplicate" review (#53)
 - 'create-review': include FAS id if can be determined (#42)
 - 'import': prompt whether to build or only push
+- 'install': add --existing-only, --skip-existing, --no-reinstall options from select-rpms-0.2 (ported from koji-tool --install)
 - 'parallel': koji-tool tail fail build.log if less than 3 packages in layer
 - 'request-repo': now offers to import the new repo immediately
 - 'request-repo': only post comment if prompt input